fix(round9): 修 CI 编译错误
- QqMusicOrder.process 重复声明 wantsCard/wantsVoice（Conflicting declarations）
- ChatAnalysisPng.drawFooter 的 textTop 由 Int 改 Float（RectF/fitBaseline 参数类型不匹配）

+++ app/src/main/java/dev/ujhhgtg/wekit/features/items/chat/ChatAnalysisPng.kt
@@ -1276,7 +1276,7 @@ object ChatAnalysisPng {
 
         // 页脚内部节奏全部来自常量：上留白 → 文字行 → 间距 → 品牌条，
         // 文字行与品牌条严格分离（避免文字下缘与渐变条压在一起）
-        val textTop = top + FOOTER_TOP_GAP
+        val textTop = (top + FOOTER_TOP_GAP).toFloat()
         val textH = FOOTER_TEXT_H.toFloat()
         val limitBottom = textTop + textH
         val pageP = paint(FS_SMALL, COLOR_ACCENT, bold = true)
